Adjust digest algorithms used in JAR signing
This fixes the issue where the digest algorithm used for APK contents
and MANIFEST.MF was unnecessarily stronger than the digest algorithm
used for PKCS #7 CMS signature. This occurred when signing
minSdkVersion 18, 19 and 20 APKs using DSA or EC keys. There, the APK
entries' contents and MANIFEST.MF were digested using SHA-256 but the
signature was generated using SHA-1 (because DSA and ECDSA signatures
with SHA-256 are supported only for minSdkVersion 21 and higher).

The fix switches to using the same digest for APK entries' contents,
MANIFEST.MF and the signature. There is no point in making some parts
of the JAR signature chain use a stronger digest than the rest.

Bug: 27461702
Change-Id: Ib300356886f54666226ae821bab89d28528f4c69