Opening a KeyStore requires a password
This removes support for opening a KeyStore without using a password.
Some KeyStore implementations use the KeyStore password only for
tamper detection of the KeyStore. Other implementations use the
KeyStore-level password to protect certificates associated with
private keys, but don't report an error when opening such a KeyStore
without a password. Thus, it's safer and cleaner to always require
a password to open a KeyStore.

Bug: 27461702
Change-Id: Ic4bd9a6e4a6614b30153b3947d3e66d8f64f7d8a