Accept non-release platform codenames in minSdkVersion
This makes apksig library and apksigner understand standard codenames
in android:minSdkVersion attribute of AndroidManifest.xml. The logic
is based on the following facts:
* standard codenames start with an upper-case A-Z letter
* API Level of platform is incremented at least by one for every
  increment of the first letter of the codename. This is useful
  for obtaining the lower bound for API Levels corresponding to
  future codenames.

This is only useful for APKs built for non-release versions of the
Android platform, such as Android Developer Preview builds. There, to
take advantage of the latest shiny new APIs, the APK's minSdkVersion
and targetSdkVersion must be set to the platform's codename, such as
"N". Such APKs are rejected by release builds of the Android platform.

Pre-release versions of the Android platform might not yet have all
the features of the final release version. Thus, apksig uses a
pessimistic esimate of a pre-release platform's API Level, usually at
least one lower than the API Level with which the platform is
eventually released. For example, non-release platforms with codename
"N" are assumed API Level 23 to be safe, rather than API Level 24 with
which Android Nougat was released publicly.

Test: bazel test :all
Bug: 33079788
Change-Id: I792df4082f84470091242910462eded53bb3ff8e