Use a more stable Created-By in .SF files
There's no need to leak information about the name and version of the
signing tool in JAR signature's .SF file. As an added benefit, signed
APKs are more reproducible because they don't change on version number
bumps.

This commit switches the default of apksig and apksigner to
"1.0 (Android)", identical to what's used by Android Plugin for Gradle
and Android Studio.

Test: apksigner sign --ks test.keystore temp.apk, then inspect .SF file
Change-Id: Ib3c742f0f15fb880afafa95de072aec16449c425