Add --in option to apksigner
This lets the user specify the APK being signed using --in option
instead of as the very last parameter on the command-line. For
symmetry reasons, apksigner verify command now also supports --in.
The original method of specifying the APK as the very last parameter
remains supported.

The reason for this change is that when --out option is used, prior to
this change, the command-line is harder to understand because the
output file precedes the input file (specified as the very last
parameter) and because it's not immediately clear from looking at the
command-line what the input file is:
  PRIOR TO THIS CHANGE
$ apksigner sign --ks .. --out app.apk app-unsigned.apk

  AFTER THIS CHANGE
$ apksigner sign --ks .. --in app-unsigned.apk --out app.apk

Test: apksigner sign --ks test.ks app.apk
Test: apksigner sign --ks test.ks --in app.apk --out app2.apk
Test: apksigner verify app.apk
Test: apksigner verify --in app.apk
Test: apksigner verify --in app.apk --print-certs
Change-Id: I753fbc234d426621717beea7179384cf15236fdf