Support keystore pwd != key pwd without extra parameters
This change makes it easier to use keystores where key password
differs from keystore password.

Prior to this change, by design, when key password spec was not
specified, key password was assumed to be the same as keystore
password. Unfortunately, this forces the user to specify --ks-pass
parameter to use such keystores. This change removes the need to do
so. When --key-pass is not provided on the command-line and keystore
password is obtained from console/stdin, the tool's behavior is now
relaxed to falling back to prompting for key password if the key
cannot be loaded using the keystore password.

https://code.google.com/p/android/issues/detail?id=233426

Test: All permutations of {keystore and key password same, different},
      {--ks-pass specified, not specified},
      {--key-pass specified, not specified}.
Bug: 34979180
Change-Id: Ic93cfff6423d25fad7103fc2ecefb0b254867d70